moved worker calculation folder into dist folder
